<REFACTOR> Logger는 앱 전반에서 쓰이므로 core.util 패키지로 이동 시킴
Logger를 공용 패키지인 core로 이동시킴으로써 타 패키지들과 interceptors 패키지의 의존성을 분리시킴.
